Add support for azurefile CSI provider
Some Azure resource identifiers don't match the format Peerpodvolume
expects, so we normalize those.

By default, `attachRequired` is false on the azurefile CSI provider.
Since the csi-wrapper assumes ControllerPublishVolume() is called, we
change `attachRequired` to true. In the future we should autodetect
whether the driver has `attachRequired` set and change the wrapper
behavour according to it.

Fixes: #585

Signed-off-by: Kartik Joshi <kartikjoshi@microsoft.com>
Signed-off-by: Iago López Galeiras <iagol@microsoft.com>

Co-authored-by: Iago López Galeiras <iagol@microsoft.com>